Move TestEnv into `given`.
